requirements.txt
altair==6.0.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
feedparser==6.0.12
fonttools==4.61.1
frozendict==2.4.7
gitdb==4.0.12
GitPython==3.1.46
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
multitasking==0.0.12
narwhals==2.17.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
peewee==4.0.1
pillow==12.1.1
platformdirs==4.9.4
protobuf==6.33.5
pyarrow==23.0.1
pycparser==3.0
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
sgmllib3k==1.0.0
soupsieve==2.8.3
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
yfinance==1.2.0
